Refactoring and adding checks.
Improved error handling and ping is now an admin/mod command.

Changed the nature of the ping command. It is essentially a message sent to a user via dm if allowed, other wise it will notify them in the appropriate channel (preferably the system channel). It's intended as a pre-warning before the bot marks them as inactive and again a few days before they are kicked for inactivity.

Any admin or mod can use this command manually. It requires that the admin create either (or both) a role named Moderator and/or Administrator role.

This is problematic as each server has their own idea on what these roles are called. Potential solution: add a global variable outside of the class. The setup command (and will also attempt to do so upon the triggering of on_guild_join in setup_tasks.py)would then ask the admin to list their moderator/admin roles and save these roles to the database for future reference.