Created local data storage.
I don't want the bot to totally rely on the database. This will provide short term storage so that an auto-save feature can be implemented to reduce the need the for database calls. It will also allow the bot to operate in the event the backend becomes unavailable, and hopefully operate quicker.